container_content/requirements.txt
aniso8601==9.0.1
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-RESTful==0.3.9
github3.py==3.2.0
h11==0.13.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
outcome==1.1.0
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-sonarqube-api==1.2.9
pytz==2022.1
PyYAML==6.0
requests==2.27.1
selenium==4.1.5
six==1.16.0
sniffio==1.2.0
sonarqube-py==0.5.4
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
uritemplate==4.1.1
urllib3==1.26.9
waitress==2.1.1
webdriver-manager==3.5.4
Werkzeug==2.1.2
wsproto==1.1.0
zipp==3.8.0
